pyproject.toml
[tool.poetry]
name = "ocr-text-blocks"
version = "0.1.0"
description = ""
authors = ["adam <adam.wk@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
